refactor: vo를 vo 패키지로 이동
