+++ src/main/java/lotto/domain/vo/Lotto.java
@@ -1,10 +1,11 @@
-package lotto.domain;
+package lotto.domain.vo;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import lotto.domain.LottoPrize;
 
 public class Lotto {
