+++ src/main/java/lotto/domain/vo/WinningNumbers.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.vo;
 
 import java.util.HashSet;
 import java.util.List;
